services/sentiment-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
redis>=5.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
yfinance>=0.2.40
tweepy>=4.14.0
praw>=7.7.0
beautifulsoup4>=4.12.0
vaderSentiment>=3.3.0
textblob>=0.17.0
openai>=1.3.0
celery>=5.3.0
schedule>=1.2.0
fake-useragent>=1.4.0
pytz>=2023.3
great-expectations>=0.18.0
scipy>=1.11.0
scikit-learn>=1.3.0
evidently>=0.4.0
transformers>=4.35.0
torch>=2.1.0
datasets>=2.14.0
accelerate>=0.24.0
evaluate>=0.4.0
seqeval>=1.2.2
wandb>=0.16.0
